requirements.txt
requests==2.32.3
PyExecJS==1.5.1
loguru==0.7.2
pycryptodome==3.20.0
openai-whisper==20231117
setuptools-rust==1.9.0
betterproto==2.0.0b7
websocket-client==1.8.0
numpy==1.26.4
sentence-transformers==3.0.1
deep_translator==1.11.4
translate==3.6.1
openai==1.37.1
opencc==1.1.9
schedule==1.2.2
selenium==4.9.0
fake-useragent==1.5.1
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
httpx==0.24.0
Pillow==9.5.0
playwright==1.42.0
tenacity==8.2.2
opencv-python
aiomysql==0.2.0
redis~=4.6.0
pydantic==2.5.2
aiofiles~=23.2.1
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
jieba==0.42.1
wordcloud==1.9.3
matplotlib==3.9.0
requests==2.32.3
parsel==1.9.1
pyexecjs==1.5.1
dataclasses==0.8